пределала последние замечания
